Some PlaceholdersService code
