+++ plugin/src/main/java/net/dzikoysk/funnyguilds/listener/AbstractFunnyListener.java
@@ -6,7 +6,7 @@ import net.dzikoysk.funnyguilds.concurrency.ConcurrencyManager;
 import net.dzikoysk.funnyguilds.config.MessageConfiguration;
 import net.dzikoysk.funnyguilds.config.PluginConfiguration;
 import net.dzikoysk.funnyguilds.config.tablist.TablistConfiguration;
-import net.dzikoysk.funnyguilds.feature.placeholders.PlaceholdersService;
+import net.dzikoysk.funnyguilds.feature.placeholders.LegacyPlaceholdersService;
 import net.dzikoysk.funnyguilds.feature.prefix.IndividualPrefixManager;
 import net.dzikoysk.funnyguilds.guild.GuildManager;
 import net.dzikoysk.funnyguilds.guild.GuildRankManager;
@@ -53,7 +53,7 @@ public abstract class AbstractFunnyListener implements Listener {
     @Inject
     public IndividualPrefixManager individualPrefixManager;
     @Inject
-    public PlaceholdersService placeholdersService;
+    public LegacyPlaceholdersService placeholdersService;
 
     @Inject
     public NmsAccessor nmsAccessor;
